% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_clean.R
\name{clean_raw_data}
\alias{clean_raw_data}
\title{Clean raw data}
\usage{
clean_raw_data(raw_data_loc = system.file("extdata", "signif.txt", package =
  "noaaQuake"))
}
\arguments{
\item{raw_data_loc}{}
}
\value{
a DT
}
\description{
function to read in raw data and clean up the following columns:
    - DATE
    - Lat & Long
    - Location
}
\examples{
\dontrun{db <- clean_raw_data()}
}
