% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_cleaning.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Data cleaning and manipulation from raw NOAA data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{a data frame of NOAA earthquake.}
}
\value{
A \code{tibble} with a new column \code{Date} but removing
    \code{Year}, \code{Mo} & \code{Dy} variables of the input data frame.
}
\description{
Data cleaning and manipulation from raw NOAA data
}
\details{
This function takes NOAA earthquake data and creates a \code{Date} column
    of type date by combining \code{Year}, \code{Mo} and \code{Dy}. It then
    makes sure that \code{Longitude} and \code{Latitude} are of type numeric
}
\examples{
data(NOAA_data)
eq_clean_data(NOAA_data)

}
