% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_graph.R
\name{eq_create_label}
\alias{eq_create_label}
\title{function to add popup text to db}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{a DT}
}
\value{
a DT
}
\description{
function to add popup text to db
}
\examples{
\dontrun{
eq_map(db[COUNTRY \%in\%c("GREECE") & DATE>"2000-01-01"],
    annot_col = "popup_text")
}
}
