% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_clean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{
a DT
}
\description{
function to clean up the location column
}
\examples{
\dontrun{db <- eq_location_clean(db)}
}
