% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_graph.R
\name{eq_map}
\alias{eq_map}
\title{Plot quakes on leaflet map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{a DT}

\item{annot_col}{string with name of annotation column}
}
\value{
a leaflet map
}
\description{
Plot quakes on leaflet map
}
\examples{
\dontrun{
eq_map(db[COUNTRY \%in\%c("GREECE") & DATE>"2000-01-01"],
    annot_col = "DATE")
}
}
