% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom function for creating the timeline from Earthquake data from NOAA}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  xmin = as.Date("0001-01-01"),
  xmax = as.Date("2020-01-01"),
  ...
)
}
\arguments{
\item{mapping}{set aesthetics}

\item{data}{The data of which we want to make the timeline}

\item{stat}{The statistics that you want to use}

\item{position}{Position of the string}

\item{na.rm}{If False, null value are removed}

\item{show.legend}{Should layer have a legend or not}

\item{inherit.aes}{Basically inherit the aesthetics or not}

\item{xmin}{Providing the minimum date for the data to be plotted}

\item{xmax}{Providing maximum date for the data to be plotted}

\item{...}{Other arguments that can be passed}
}
\value{

}
\description{
Geom function for creating the timeline from Earthquake data from NOAA
}
\examples{
\dontrun{
data(NOAA_data)
NOAA_data \%>\% eq_clean_data() \%>\%
dplyr::filter(Country \%in\% c("India", "Pakistan")) \%>\%
   ggplot(aes(x = date, y = Country, color = Total.Deaths, size = Mag,
              magnitude = Mag, label =Location_Name)) +
   labs(color="#deaths",size="Magnitude",x="DATE")+
   geom_timeline(alpha=0.2,xmin = ymd('2000-01-01'), xmax = ymd('2015-12-31'))+
  labs(x="DATE")+theme_minimal()
}
}
