% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Adding label to Geom Timeline}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{set aesthetics}

\item{data}{provide the dataframe of the labels}

\item{stat}{Any statistics}

\item{position}{Position of the label}

\item{...}{Other arguments that can be passed}

\item{na.rm}{TRUE if you want to remove the missing and False if not}

\item{show.legend}{TRUE to show legend}

\item{inherit.aes}{Basically inherit the aesthetics or not}
}
\value{
This returns the geom timeline with label
}
\description{
Adding label to Geom Timeline
}
\examples{
\dontrun{
labels <- NOAA_data \%>\% eq_clean_data() \%>\%
dplyr::filter(Country \%in\% c("India", "Pakistan"))\%>\%
   filter(date> ymd('2000-01-01') & date <ymd('2015-12-31'))\%>\%
  top_n(5,Mag)

NOAA_data \%>\% eq_clean_data() \%>\%
   dplyr::filter(Country \%in\% c("Pakistan", "India")) \%>\%
   ggplot(aes(x = date, y = Country, color = Total.Deaths, size = Mag,
              magnitude = Mag, label =Location_Name)) +
   geom_timeline(alpha=0.4,xmin = ymd('2000-01-01'), xmax = ymd('2015-12-31'))+
   geom_timeline_label(data = labels,
                       mapping = aes(x = date, y = Country, label = Location_Name))+
   labs(color="#deaths",size="Richter scale value",x="DATE")+
   theme_minimal()
   }
}
