% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{ggproto timeline object of class ggproto}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
GeomTimeline
}
\description{
This is the setup required for creating a new geom class. This GeomTimeLine
    inherits from a top level class called Geom.It creates a geom that draws the
    timeline for a specified date interval and put points on it for each earthquake.
}
\keyword{datasets}
