% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{ggproto GeomTimeline Label object of class ggproto}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimelineLabel
}
\description{
This is the setup required for creating a new geom class. This GeomTimeLabel
    inherits from a top level class called Geom.It creates a geom that draws the
    label of the eartquae for a specified date interval and put points on it for each earthquake.
}
\keyword{datasets}
