% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NOAA_data}
\alias{NOAA_data}
\title{NOAA Significant Earthquake Database}
\format{
A data frame of 475 observations and 39 variables.
}
\source{
NOAA National Centers for Environmental Information
 National Geophysical Data Center / World Data Service (NGDC/WDS):
 NCEI/WDS Global Significant Earthquake Database. NOAA National Centers
 for Environmental Information.
 doi:10.7289/V5TD9V7K
 \url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}
}
\usage{
NOAA_data
}
\description{
This dataset was obtained from U.S. National Oceanographic and Atmospheric
 Administration (NOAA) on significant earthquakes around the world. It
 contains information about 475 earthquakes over an
 approximately 4,000 year time span in India and Pakistan.
}
\keyword{datasets}
