% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Adding customized label in the Leaflet map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Adding the data dataframe}
}
\value{
This returns a HTML tags for the customized label
}
\description{
Adding customized label in the Leaflet map
}
\examples{
\dontrun{
data_clean <- NOAA_data \%>\%
eq_clean_data()\%>\%
   dplyr::filter(Country == "India"|Country == "Pakistan" & lubridate::year(date) >= 2000)\%>\%
   dplyr::mutate(popup_text = eq_create_label(.))
eq_map(data = data_clean, annot_col = "popup_text")
}
}
