% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_map.R
\name{eq_map}
\alias{eq_map}
\title{Leaflet map tool}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{Adding the dataframe}

\item{annot_col}{Annotate column that will come as a popup}
}
\value{
It returns a leaflet map
}
\description{
Leaflet map tool
}
\examples{
\dontrun{
data_clean <- NOAA_data \%>\%
eq_clean_data() \%>\%
dplyr::filter(Country == "India"|Country == "Pakistan" & lubridate::year(date) >= 2000)
eq_map(data = data_clean, annot_col = "Mag")
}
}
